requirements.txt
groq
python-telegram-bot
PyPDF2
python-dotenv
python-docx
reportlab
